+++ CHANGELOG.md
@@ -10,6 +10,6 @@
     * First stack damage bonus: 20% -> 15%
     * Barrier out-of-combat delay time: 3s -> 5s
     * Barrier out-of-combat gain: 6% -> 8%
+* Whimsical Star : Is now achievement locked
 * Added Relic Guard
 * Added Super Shield (temp name - also functionality changed from doc)
-* Added Heal Exp (temp name)